scraper/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
psycopg2-binary==2.9.9
python-decouple==3.8
lxml==4.9.3
fake-useragent==1.4.0
schedule==1.2.0
urllib3==2.0.7